UX개선 : 사진 업로드 시간이 길어질 때, 사용자가 이미지 업로드 완료 여부를 파악하기 힘든 문제 개선 (#667)
* feature: 이미지 업로드 시 버튼 조건부 disable기능 추가

* chore: 컨벤션 준수를 위한 코드 수정

* refactor: isImageUploading상태 네이밍 수정

* feat: 프로필 사진 업로드 시 버튼 disabled 기능 추가

* chore: 불필요한 storybook파일 제거
