+++ frontend/src/components/myPage/EditUserProfileForm/EditUserProfileForm.tsx
@@ -1,15 +1,21 @@
+import { useSingleImageUpload } from '@/hooks/common/useSingleImageUpload';
+
+import { useCallback, useRef } from 'react';
+
 import { Box, Button, Flex, Heading, Modal, Text, useOverlay } from 'hang-log-design-system';
 
 import {
   buttonStyling,
   deleteButtonStyling,
   formStyling,
-  imageInputStyling,
+  imageStyling,
+  inputStyling,
   modalButtonContainerStyling,
   modalContentStyling,
+  uploadButtonStyling,
+  wrapperStyling,
 } from '@components/myPage/EditUserProfileForm/EditUserProfileForm.style';
 import NicknameInput from '@components/myPage/EditUserProfileForm/NicknameInput/NicknameInput';
-import ProfileImageInput from '@components/myPage/EditUserProfileForm/ProfileImageInput/ProfileImageInput';
 
 import { useDeleteAccountMutation } from '@hooks/api/useDeleteAccountMutation';
 import { useEditUserProfileForm } from '@hooks/member/useEditUserProfileForm';
@@ -21,6 +27,8 @@ interface EditUserProfileForm {
 }
 
 const EditUserProfileForm = ({ initialData }: EditUserProfileForm) => {
+  const inputRef = useRef<HTMLInputElement | null>(null);
+
   const { userInfo, isNicknameError, updateInputValue, disableNicknameError, handleSubmit } =
     useEditUserProfileForm(initialData);
 
@@ -36,21 +44,56 @@ const EditUserProfileForm = ({ initialData }: EditUserProfileForm) => {
     deleteAccountMutation.mutate();
   };
 
+  const handleImageUploadButtonClick = () => {
+    inputRef.current?.click();
+  };
+
+  const handleImageUrlsChange = useCallback(
+    (imageUrl: string) => {
+      updateInputValue('imageUrl', imageUrl);
+    },
+    [updateInputValue]
+  );
+
+  const { isImageUploading, uploadedImageUrl, handleImageUpload } = useSingleImageUpload({
+    initialImageUrl: initialData.imageUrl,
+    onSuccess: handleImageUrlsChange,
+  });
+
   return (
     <>
       <form css={formStyling} onSubmit={handleSubmit} noValidate>
-        <ProfileImageInput
-          css={imageInputStyling}
-          initialImageUrl={initialData.imageUrl}
-          updateInputValue={updateInputValue}
-        />
+        <Box css={wrapperStyling}>
+          <Button
+            css={uploadButtonStyling}
+            type="button"
+            size="small"
+            variant="text"
+            onClick={handleImageUploadButtonClick}
+          >
+            수정
+          </Button>
+          <input
+            css={inputStyling}
+            type="file"
+            accept="image/*"
+            id="profile-image"
+            ref={inputRef}
+            onChange={handleImageUpload}
+          />
+          <img css={imageStyling} src={uploadedImageUrl!} alt="사용자 프로필 이미지" />
+        </Box>
         <NicknameInput
           value={userInfo.nickname}
           isError={isNicknameError}
           updateInputValue={updateInputValue}
           disableError={disableNicknameError}
         />
-        <Button variant="primary" css={buttonStyling}>
+        <Button
+          variant={isImageUploading ? 'default' : 'primary'}
+          css={buttonStyling}
+          disabled={isImageUploading}
+        >
           수정하기
         </Button>
         <Button
