+++ frontend/src/hooks/api/useCommunityTripQuery.ts
@@ -1,10 +1,10 @@
-import { isLoggedInState } from '@/store/auth';
-
 import { useQuery } from '@tanstack/react-query';
 
 import { useRecoilValue } from 'recoil';
 
 import type { AxiosError } from 'axios';
+
+import { isLoggedInState } from '@store/auth';
 
 import { getCommunityTrip } from '@api/trip/getCommunityTrip';
