+++ frontend/src/hooks/api/useTripQuery.ts
@@ -1,17 +1,18 @@
-import { TRIP_TYPE } from '@/constants/trip';
-import { isLoggedInState } from '@/store/auth';
-
 import { useQuery } from '@tanstack/react-query';
 
 import { useRecoilValue } from 'recoil';
 
 import type { AxiosError } from 'axios';
 
+import { isLoggedInState } from '@store/auth';
+
 import { getCommunityTrip } from '@api/trip/getCommunityTrip';
 import { getSharedTrip } from '@api/trip/getSharedTrip';
 import { getTrip } from '@api/trip/getTrip';
 
 import type { TripData, TripTypeData } from '@type/trip';
+
+import { TRIP_TYPE } from '@constants/trip';
 
 export const useTripQuery = (tripType: TripTypeData, tripId: string) => {
   const isLoggedIn = useRecoilValue(isLoggedInState);
